Move session_counts_reset() to utils.py and update firestore code to load/save session separately from agg counts
Had to do the function move due to circular import issues between firestore.py and main.py
