+++ src/streamlit_analytics2/utils.py
@@ -1,3 +1,7 @@
+import datetime
+from typing import Any, Dict
+
+
 def format_seconds(s: int) -> str:
     """Formats seconds to 00:00:00 format."""
     # days, remainder = divmod(s, 86400)
@@ -26,3 +30,25 @@ def replace_empty(s):
         return " "
     else:
         return s
+    
+    
+def session_counts_reset() -> Dict[str, Any]:
+    """
+    Reset the session counts to a new session.
+
+    Returns
+    -------
+    Dict[str, Any]
+        The new session counts.
+    """
+    # Use yesterday as first entry to make chart look better.
+    yesterday = str(datetime.date.today() - datetime.timedelta(days=1))
+    output = {}
+    output["total_pageviews"] = 0
+    output["total_script_runs"] = 0
+    output["total_time_seconds"] = 0
+    output["per_day"] = {"days": [str(yesterday)], "pageviews": [0], "script_runs": [0]}
+    output["widgets"] = {}
+    output["start_time"] = datetime.datetime.now().strftime("%d %b %Y, %H:%M:%S")
+    
+    return output
